+++ src/Summoner/Template.hs
@@ -25,30 +25,32 @@ import qualified Data.Text as T
 -- Stack File Creation
 ----------------------------------------------------------------------------
 
-emptyIfNot :: Monoid m => Bool -> m -> m
-emptyIfNot p val = if p then val else mempty
+memptyIfFalse :: Monoid m => Bool -> m -> m
+memptyIfFalse p val = if p then val else mempty
 
 -- | Creating template file to use in `stack new` command
 createStackTemplate :: ProjectData ->  TreeFs
 createStackTemplate ProjectData{..} = Dir (toString repo) $
     [ File (toString repo <> ".cabal")
            ( createCabalTop
-          <> emptyIfNot github createCabalGit
-          <> emptyIfNot isLib createCabalLib
-          <> emptyIfNot isExe
+          <> memptyIfFalse github createCabalGit
+          <> memptyIfFalse isLib createCabalLib
+          <> memptyIfFalse isExe
                         ( createCabalExe
-                        $ emptyIfNot isLib $ ", " <> repo )
-          <> emptyIfNot test createCabalTest
-          <> emptyIfNot bench
+                        $ memptyIfFalse isLib $ ", " <> repo )
+          <> memptyIfFalse test
+                        ( createCabalTest
+                        $ memptyIfFalse isLib $ ", " <> repo )
+          <> memptyIfFalse bench
                         ( createCabalBenchmark
-                        $ emptyIfNot isLib $ ", " <> repo )
+                        $ memptyIfFalse isLib $ ", " <> repo )
            )
     , File "README.md" readme
     , File "CHANGELOG.md" changelog
     , File "LICENSE" licenseText
     ]
  ++ createCabalFiles
- ++ emptyIfNot stack (createStackYamls testedVersions)
+ ++ memptyIfFalse stack (createStackYamls testedVersions)
  ++ [File ".gitignore" gitignore | github]
  ++ [File ".travis.yml" travisYml | travis]
  ++ [File "appveyor.yml" appVeyorYml | appVey]
@@ -109,6 +111,36 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
         $endLine
         |]
 
+    ghcOptions :: Text
+    ghcOptions = case warnings of
+        [] -> defaultWarnings
+        xs -> T.intercalate "\n" xs
+
+    defaultWarnings :: Text
+    defaultWarnings =
+        [text|
+        -Wincomplete-uni-patterns
+        -Wincomplete-record-updates
+        -Wmissing-import-lists
+        -Wcompat
+        -Widentities
+        $versionWarnings
+        |]
+
+    versionWarnings :: Text
+    versionWarnings
+        =  memptyIfFalse (testedVersions `hasLeast` Ghc801)
+            "-Wredundant-constraints\n"
+        <> memptyIfFalse (testedVersions `hasLeast` Ghc822)
+            "-fhide-source-paths\n"
+        <> memptyIfFalse (testedVersions `hasLeast` Ghc843)
+            [text|
+            -Wmissing-export-lists
+            -Wpartial-fields
+            |]
+      where
+        hasLeast list el = all (>= el) list
+
     createCabalLib :: Text
     createCabalLib =
         [text|
@@ -117,6 +149,7 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
           exposed-modules:     $libModuleName
                                $preludeMod
           ghc-options:         -Wall
+                               $ghcOptions
           build-depends:       $base
                              $customPreludePack
           default-language:    Haskell2010
@@ -130,7 +163,11 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
         executable $repo
           hs-source-dirs:      app
           main-is:             Main.hs
-          ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N
+          ghc-options:         -Wall
+                               -threaded
+                               -rtsopts
+                               -with-rtsopts=-N
+                               $ghcOptions
           build-depends:       $base
                              $r
                              $customPreludePack
@@ -139,17 +176,21 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
         $endLine
         |]
 
-    createCabalTest :: Text
-    createCabalTest =
+    createCabalTest :: Text -> Text
+    createCabalTest r =
         [text|
         test-suite ${repo}-test
           type:                exitcode-stdio-1.0
           hs-source-dirs:      test
           main-is:             Spec.hs
+          ghc-options:         -Wall
+                               -threaded
+                               -rtsopts
+                               -with-rtsopts=-N
+                               $ghcOptions
           build-depends:       $base
-                             , $repo
+                             $r
                              $customPreludePack
-          ghc-options:         -Wall -Werror -threaded -rtsopts -with-rtsopts=-N
           default-language:    Haskell2010
           $defaultExtensions
         $endLine
@@ -159,15 +200,20 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
     createCabalBenchmark r =
         [text|
         benchmark ${repo}-benchmark
-          type:                exitcode-stdio-1.0
-          default-language:    Haskell2010
-          ghc-options:         -Wall -Werror -O2 -threaded -rtsopts -with-rtsopts=-N
-          hs-source-dirs:      benchmark
-          main-is:             Main.hs
-          build-depends:       $base
-                             , gauge
-                             $customPreludePack
-                             $r
+          type:               exitcode-stdio-1.0
+          hs-source-dirs:     benchmark
+          main-is:            Main.hs
+          ghc-options:        -Wall
+                              -threaded
+                              -rtsopts
+                              -with-rtsopts=-N
+                              -02
+                              $ghcOptions
+          build-depends:      $base
+                            , gauge
+                            $r
+                            $customPreludePack
+          default-language:   Haskell2010
           $defaultExtensions
         $endLine
         |]
@@ -280,7 +326,7 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
         stackLinkNightly :: Text =
             "http://stackage.org/nightly/package/" <> repo
 
-        stackBadges :: Text = emptyIfNot stack
+        stackBadges :: Text = memptyIfFalse stack
             [text|
             [![Stackage Lts](${stackShieldLts})](${stackLinkLts})
             [![Stackage Nightly](${stackShieldNightly})](${stackLinkNightly})
@@ -290,14 +336,14 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
           "https://secure.travis-ci.org/" <> owner <> "/" <> repo <> ".svg"
         travisLink :: Text =
           "https://travis-ci.org/" <> owner <> "/" <> repo
-        travisBadge :: Text = emptyIfNot travis
+        travisBadge :: Text = memptyIfFalse travis
             [text|[![Build status](${travisShield})](${travisLink})|]
 
         appVeyorShield :: Text =
           "https://ci.appveyor.com/api/projects/status/github/" <> owner <> "/" <> repo <> "?branch=master&svg=true"
         appVeyorLink :: Text =
           "https://ci.appveyor.com/project/" <> owner <> "/" <> repo
-        appVeyorBadge :: Text = emptyIfNot appVey
+        appVeyorBadge :: Text = memptyIfFalse appVey
             [text|[![Windows build status](${appVeyorShield})](${appVeyorLink})|]
 
         licenseShield :: Text =
@@ -387,10 +433,10 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
     -- create travis.yml template
     travisYml :: Text
     travisYml =
-        let travisStackMtr = emptyIfNot stack $
+        let travisStackMtr = memptyIfFalse stack $
                 T.concat (map travisStackMatrixItem $ delete defaultGHC testedVersions)
                     <> travisStackMatrixDefaultItem
-            travisCabalMtr = emptyIfNot cabal $
+            travisCabalMtr = memptyIfFalse cabal $
                 T.concat $ map travisCabalMatrixItem testedVersions
             installAndScript =
                 if cabal
@@ -398,8 +444,8 @@ createStackTemplate ProjectData{..} = Dir (toString repo) $
                      then installScriptBoth
                      else installScriptCabal
                 else installScriptStack
-            travisCabalCache = emptyIfNot cabal "- \"$HOME/.cabal\""
-            travisStackCache = emptyIfNot stack
+            travisCabalCache = memptyIfFalse cabal "- \"$HOME/.cabal\""
+            travisStackCache = memptyIfFalse stack
                 [text|
                 - "$$HOME/.stack"
                 - "$$TRAVIS_BUILD_DIR/.stack-work"