+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   Step into created project properly
 * [#110](https://github.com/kowainik/summoner/issues/110):
   Print executing command
+* [#91](https://github.com/kowainik/summoner/issues/91):
+  Add warning fields for ghc
 
 1.0.6
 =====